ic_sm/scripts_4_db/requirements.txt
cryptography
psycopg2-binary
